Move hackathon contra/nesci work to contra-demo subproject.
Note: nothing installs the pycontra python module yet, so
the python client won't run out of the box.

* Set up contra-demo as external project.
* Remove dependencies on arbor's internal aux/anc library.
* Strip out MPI, profile, meter, json code (MPI not supported in the
demo yet anyway).
* Manually add contra-config.cmake file.
