--- example/risotto/CMakeLists.txt
@@ -1,17 +1,0 @@
-find_package(contra COMPONENTS core zmq QUIET)
-find_package(nesci COMPONENTS layout producer QUIET)
-
-if (NOT contra_FOUND)
-    message(STATUS "Optional dependency 'contra' not found, skipping creation of example 'risotto'")
-elseif (NOT nesci_FOUND)
-    message(STATUS "Optional dependency 'nesci' not found, skipping creation of example 'risotto'")
-else ()
-    add_executable(risotto risotto.cpp)
-    target_link_libraries(risotto PRIVATE arbor arbor-anc ext-tclap ext-json contra contra_zmq nesci_producer)
-    add_custom_command(
-        TARGET risotto POST_BUILD
-        COMMAND ${CMAKE_COMMAND}
-        -E copy ${CMAKE_CURRENT_SOURCE_DIR}/risotto_client.py ${CMAKE_BINARY_DIR}/bin/risotto_client.py
-        DEPENDS ${CMAKE_CURRENT_SOURCE_DIR}/risotto_client.py
-    )
-endif ()
