--- example/risotto/readme.md
@@ -1,3 +1,0 @@
-# Ring Example
-
-A miniapp that demonstrates how to describe how to build a simple ring network.
